Add note that replay buffer size is per env (#1)
*Description of changes:*

Tiny update: add buffer_size docstring in FastSACConfig to state it is
per environment.

Clarifies the meaning of buffer_size in FastSACConfig. Currently the
docstring suggests it is a single replay buffer size, but per
[fast_sac_utils.py#L42](https://github.com/amazon-far/holosoma/blob/d52658839e9a904cf3dc526cbc854cec589cc447/src/holosoma/holosoma/agents/fast_sac/fast_sac_utils.py#L42),
the buffer is allocated per environment.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Cole Harrison <colha@amazon.com>